requirements.txt
python-dotenv
requests
lxml
loguru
psycopg2-binary
pandas 
openpyxl 
unidecode